Enhance questions form
