+++ src/commonComponents/AddOrEditQuiz/Components/AnswersFormSection/Components/CorrectAnswer/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CorrectAnswer";
