+++ src/commonComponents/AddOrEditQuiz/Components/AnswersFormSection/index.js
@@ -1,0 +1,1 @@
+export { default } from "./AnswersFormSections";
