+++ src/commonComponents/AddOrEditQuiz/Components/QuestionsFormSection/QuestionsFormSection.js
@@ -1,31 +1,13 @@
-import { Box, Button, Grid, Typography } from "@mui/material";
+import { Button, Grid } from "@mui/material";
 import { useFormikContext } from "formik";
 import { FieldArray } from "formik";
 import FormField from "../../../FormField/FormField";
 import RemoveCircleOutlineIcon from "@mui/icons-material/RemoveCircleOutline";
 import AddCircleOutlineIcon from "@material-ui/icons/AddCircleOutline";
-import { FormControlLabel, Radio } from "@mui/material";
+import AnswersFormSection from "../AnswersFormSection";
 
 const QuestionsFormSection = () => {
-  const { values, setFieldValue } = useFormikContext();
-
-  const handleSelectAnswer = (questionIndex, answerIndex) => {
-    // Fuction to handle that only one answer is selected per question and updating is_true value
-    const updatedAnswers = values.questions_answers.map((question, qIndex) => {
-      if (qIndex === questionIndex) {
-        return {
-          ...question,
-          answers: question.answers.map((answer, aIndex) => ({
-            ...answer,
-            is_true: aIndex === answerIndex,
-          })),
-        };
-      }
-      return question;
-    });
-
-    setFieldValue("questions_answers", updatedAnswers);
-  };
+  const { values } = useFormikContext();
 
   return (
     <FieldArray name="questions_answers">
@@ -53,6 +35,7 @@ const QuestionsFormSection = () => {
                 }}
               >
                 {values.questions_answers.length > 1 && (
+                  // To ensure that every quiz has atleast on question
                   <Button
                     style={{
                       color: "black",
@@ -107,91 +90,7 @@ const QuestionsFormSection = () => {
                 showErrorMessage
               />
             </Grid>
-            <FieldArray name={`questions_answers.${index}.answers`}>
-              {({ push, remove }) => (
-                <div>
-                  {question.answers.map((answer, answerIndex) => (
-                    <Grid container spacing={2} key={answerIndex}>
-                      <Grid item xs={8}>
-                        <FormField
-                          type="input"
-                          name={`questions_answers.${index}.answers.${answerIndex}.text`}
-                          label={`Answer ${answerIndex + 1}`}
-                          variant="standard"
-                          fullwidth
-                          showErrorMessage
-                        />
-                      </Grid>
-                      <Grid
-                        item
-                        xs={2}
-                        style={{
-                          display: "flex",
-                          justifyContet: "center",
-                          alignItems: "center",
-                        }}
-                      >
-                        {question.answers.length > 1 && (
-                          <Button
-                            style={{
-                              color: "black",
-                            }}
-                            onClick={() => remove(answerIndex)}
-                          >
-                            <RemoveCircleOutlineIcon />
-                          </Button>
-                        )}
-                      </Grid>
-                      <Grid
-                        item
-                        xs={2}
-                        style={{
-                          display: "flex",
-                          justifyContet: "center",
-                          alignItems: "center",
-                        }}
-                      >
-                        <Button
-                          style={{
-                            color: "black",
-                          }}
-                          onClick={() => push({ text: "", is_true: false })}
-                        >
-                          <AddCircleOutlineIcon />
-                        </Button>
-                      </Grid>
-                    </Grid>
-                  ))}
-                </div>
-              )}
-            </FieldArray>
-
-            <Typography
-              style={{ fontSize: 17, paddingTop: 10, paddingBottom: 10 }}
-            >
-              Correct Answer for question {index + 1}:
-            </Typography>
-            <FieldArray name={`questions_answers.${index}.answers`}>
-              {() => (
-                <Box style={{ display: "flex", flexDirection: "column" }}>
-                  {question.answers.map(
-                    (answer, answerIndex) =>
-                      answer.text && (
-                        <FormControlLabel
-                          key={answerIndex}
-                          value={answer.text}
-                          control={<Radio />}
-                          label={answer.text}
-                          checked={answer.is_true}
-                          onChange={() =>
-                            handleSelectAnswer(index, answerIndex)
-                          }
-                        />
-                      )
-                  )}
-                </Box>
-              )}
-            </FieldArray>
+            <AnswersFormSection index={index} question={question} />
           </div>
         ))
       }
